backend/requirements.txt
asgiref==3.8.1
pi==0.1.2
pip==23.2.1
psycopg==3.2.6
psycopg-binary==3.2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
pillow
atlassian-python-api==4.0.4
langchain==0.3.25
langchain-huggingface==0.2.0
langchain-community==0.3.24
unstructured==0.17.2
faiss-cpu==1.11.0
huggingface_hub[hf_xet]
markdown==3.8
torch
sentence-transformers
transformers

Django==5.1.7
djangorestframework==3.15.2
django-environ==0.12.0
requests==2.31.0
dotenv
python-docx==1.1.2
psycopg2
qdrant-client==1.15.1
langchain-core==0.3.79
langchain-gigachat==0.3.12
